verificar se esta com plano assinado

+++ lib/actions/law/law.delete.action.ts
@@ -2,9 +2,19 @@
 
 import { withPostAuth } from "@/lib/auth";
 import prisma from "@/lib/configs/prisma";
+import { hasSubscription } from "@/lib/helpers/billing";
 import { Law } from "@prisma/client";
 
 export const deleteLaw = withPostAuth(async (_: FormData, law: Law) => {
+
+  const hasSub = await hasSubscription();
+
+  if (!hasSub) {
+    return {
+      error: `Você precisa assinar um plano para realizar este comando.`,
+    };
+  }
+
   try {
     const response = await prisma.law.delete({
       where: {
