+++ lib/actions/post/post.delete.action.ts
@@ -2,10 +2,20 @@
 
 import { withPostAuth } from "@/lib/auth";
 import prisma from "@/lib/configs/prisma";
+import { hasSubscription } from "@/lib/helpers/billing";
 import { Post } from "@prisma/client";
 
 export const deletePost = withPostAuth(async (_: FormData, post: Post) => {
   try {
+
+    const hasSub = await hasSubscription();
+
+    if (!hasSub) {
+      return {
+        error: `Você precisa assinar um plano para realizar este comando.`,
+      };
+    }
+
     const response = await prisma.post.delete({
       where: {
         id: post.id,
